pyproject.toml
[project]
name = "modern-route-full-stack-generativeai-and-agentic-ai-bootcamp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=1.14.2",
    "crewai-tools>=0.76.0",
    "datasets>=2.19.0",
    "gensim==4.3.2",
    "ipykernel>=7.2.0",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.12",
    "matplotlib>=3.10.8",
    "mteb>=1.12.5",
    "nltk>=3.9.1",
    "numpy==1.26.4",
    "pandas>=2.3.3",
    "pillow>=10.2.0",
    "python-dotenv>=1.1.1",
    "rank-bm25==0.2.2",
    "scikit-learn==1.4.2",
    "scipy==1.11.4",
    "sentence-transformers==2.7.0",
    "streamlit>=1.44.0",
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "transformers>=4.41.0",
]
